fix: moved hex_serde into a new file
