+++ crates/crypto/keystore/src/encrypted_keystore.rs
@@ -1,8 +1,9 @@
 use std::{fs, path::Path};
-
 use anyhow::Result;
 use ream_bls::PubKey;
 use serde::{Deserialize, Serialize};
+
+use crate::hex_serde;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct EncryptedKeystore {
@@ -74,25 +75,6 @@ pub enum CipherParams {
 #[serde(tag = "function", content = "params", rename_all = "lowercase")]
 pub enum ChecksumParams {
     Sha256 {},
-}
-
-mod hex_serde {
-    use alloy_primitives::hex::{FromHex, ToHexExt};
-    use serde::{Deserialize, Deserializer, Serializer, de};
-    pub fn serialize<T, S>(value: &T, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        T: AsRef<[u8]>,
-        S: Serializer,
-    {
-        serializer.serialize_str(&value.encode_hex())
-    }
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(deserializer)?;
-        Vec::<u8>::from_hex(&s).map_err(|err| de::Error::custom(err.to_string()))
-    }
 }
 
 #[cfg(test)]
