+++ crates/crypto/keystore/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod encrypted_keystore;
+pub mod hex_serde;
